+++ AdvanceTaskMarsPart2/Hooks/ExtentHooks.cs
@@ -9,6 +9,7 @@ using TechTalk.SpecFlow;
 using AventStack.ExtentReports.Gherkin.Model;
 using NUnit.Framework;
 using AdvanceTaskMarsPart2.Utilities;
+using OpenQA.Selenium;
 
 namespace AdvanceTaskMarsPart2.Hooks
 {
@@ -59,6 +60,26 @@ namespace AdvanceTaskMarsPart2.Hooks
             if (scenarioContext.TestError != null)
             {
                 test.Fail("Scenario failed: " + scenarioContext.TestError.Message);
+                try
+                {
+                    var screenshot = ((ITakesScreenshot)driver).GetScreenshot();
+
+                    string screenshotDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Screenshots");
+                    if (!Directory.Exists(screenshotDirectory))
+                        Directory.CreateDirectory(screenshotDirectory);
+
+                    string safeTestName = string.Join("_", TestContext.CurrentContext.Test.Name.Split(Path.GetInvalidFileNameChars()));
+                    string screenshotFile = $"{safeTestName}_{DateTime.Now:yyyyMMdd_HHmmss}.png";
+                    string screenshotPath = Path.Combine(screenshotDirectory, screenshotFile);
+
+                    screenshot.SaveAsFile(screenshotPath);
+                    test.Fail("Test failed").AddScreenCaptureFromPath(screenshotPath);
+                }
+                catch (Exception ex)
+                {
+                    test.Fail("Failed to capture screenshot: " + ex.Message);
+                }
+
             }
             else
             {
