+++ AdvanceTaskMarsPart2/Hooks/Hooks.cs
@@ -27,34 +27,6 @@ namespace AdvanceTaskMarsPart2.Hooks
         [AfterScenario]
         public void AfterScenario()
         {
-           /* var outcome = TestContext.CurrentContext.Result.Outcome.Status;
-
-            if (outcome == TestStatus.Failed)
-            {
-                try
-                {
-                    var screenshot = ((ITakesScreenshot)driver).GetScreenshot();
-
-                    string screenshotDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Screenshots");
-                    if (!Directory.Exists(screenshotDirectory))
-                        Directory.CreateDirectory(screenshotDirectory);
-
-                    string safeTestName = string.Join("_", TestContext.CurrentContext.Test.Name.Split(Path.GetInvalidFileNameChars()));
-                    string screenshotFile = $"{safeTestName}_{DateTime.Now:yyyyMMdd_HHmmss}.png";
-                    string screenshotPath = Path.Combine(screenshotDirectory, screenshotFile);
-
-                    screenshot.SaveAsFile(screenshotPath);
-                    test.Fail("Test failed").AddScreenCaptureFromPath(screenshotPath);
-                }
-                catch (Exception ex)
-                {
-                    test.Fail("Failed to capture screenshot: " + ex.Message);
-                }
-            }
-            else
-            {
-                test.Pass("Test passed");
-            } */
 
             TearDownAction();
         }
