+++ AdvanceTaskMarsPart2/Utilities/ToastMessageHelper.cs
@@ -12,11 +12,7 @@ namespace AdvanceTaskMarsPart2.Utilities
     {
         private static readonly By toastMessageLocator = By.XPath("//div[@class='ns-box-inner']");
 
-        /// <summary>
-        /// Waits for toast message, gets its text, and verifies it matches any of the expected messages.
-        /// </summary>
-        /// <param name="expectedMessages">List of acceptable toast messages.</param>
-        /// <returns>The actual toast message text.</returns>
+       
         public static string VerifyToastMessage(IWebDriver driver, params string[] expectedMessages)
         {
             try
@@ -27,7 +23,7 @@ namespace AdvanceTaskMarsPart2.Utilities
                 string displayedMessage = toastMessage.Text;
                 Console.WriteLine("Toast message: " + displayedMessage);
 
-                // Validate against expected list
+                
                 Assert.That(displayedMessage, Is.AnyOf(expectedMessages));
 
                 return displayedMessage;
